little reorganization
