+++ .env.example
@@ -9,5 +9,4 @@ PANEL_BASE_URL="https://panel.example.org/"
 SERVER_APPLICATION_ID="anyNumberButProbably1or2or3or4or5"
 SERVER_CLIENT_ID="dj28041orSomethingLikeThat"
 
-CEDMOD_BASE_URL="https://cedmod.example.org"
 CEDMOD_INSTANCE_ID="1234"
